Adding all the files from Coventry Uni Github
For my final year project I decided to code a script that could detect possible trojan viruses from image files as well as possible macro viruses from word, excel files and executable files. In this project the user was able to upload their files on an online URL server and then see the results on terminal on whether or not their file was infected.